+++ src/title/index.tsx
@@ -1,28 +1,15 @@
-import React, { Fragment } from 'react'
-import cc from 'classcat'
-import style from 'title/style'
+import styled from 'styled-components'
+import { color, font } from '_utils/branding'
 
-interface TitleProps {
-  readonly className?: Classcat.Class
-  readonly children: string
-  readonly headingLevel?: number
-}
+import Title from './Title'
 
-export const isHeadingAvailable = (level: number) => level >= 1 && level <= 6
-
-const Title = ({ className, children, headingLevel = 1 }: TitleProps) => {
-  if (!isHeadingAvailable(headingLevel)) {
-    return null
+const StyledTitle = styled(Title)`
+  & {
+    color: ${color.primaryText};
+    font-size: ${font.xl.size};
+    font-weight: 500;
+    line-height: ${font.xl.lineHeight};
   }
-  return React.createElement(
-    `h${headingLevel}`,
-    {
-      className: cc(['kirk-title', className]),
-    },
-    <Fragment>
-      {children}
-      <style jsx>{style}</style>
-    </Fragment>,
-  )
-}
-export default Title
+`
+
+export default StyledTitle
